+++ Source/Components/Touchpoints/MLBusinessImageProvider.swift
@@ -1,0 +1,12 @@
+//
+//  MLBusinessImageProvider.swift
+//  MLBusinessComponents
+//
+//  Created by Vicente Veltri on 07/07/2020.
+//
+
+import Foundation
+
+public protocol MLBusinessImageProvider: class {
+    func getImage(key: String, completion: @escaping (UIImage?) -> Void)
+}