+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-# Sin publicar
+## Sin Publicar
 - Changed Dividing Line View stroke color (darker)
+- Make MLBusinessTouchpointView Carousel public
 
 # v1.15.0
 🚀 1.15.0 🚀